requirements.txt
accelerate==1.1.0
compressai==1.2.8
diffusers==0.35.2
einops==0.8.2
lpips==0.1.4
numpy==2.4.3
openai_clip==1.0.1
peft==0.17.1
Pillow==12.1.1
pyiqa==0.1.14.1
PyYAML==6.0.3
taming_transformers==0.0.1
timm==1.0.25
torch==2.8.0+cu128
torch_fidelity==0.3.0
torchmetrics==1.8.2
torchvision==0.23.0+cu128
tqdm==4.67.1
xformers==0.0.35
